refactoring based on discord discussion
